🧹 refactor config gathering for sbom cmd
